Convert H5Z*.jl packages into package extensions
