+++ ext/CodecBzip2Ext/README.md
@@ -1,0 +1,6 @@
+# H5Zbzip2.jl
+
+Implements the Bzip2 filter for [HDF5.jl](https://github.com/JuliaIO/HDF5.jl) in Julia.
+See the [documentation](https://juliaio.github.io/HDF5.jl/stable/filters/#H5Zbzip2.jl)
+
+This implements [HDF5 registered filter id 307](https://portal.hdfgroup.org/display/support/Filters#Filters-307).
