+++ ext/CodecLz4Ext/README.md
@@ -1,0 +1,6 @@
+# CodecLz4Ext
+
+Implements the LZ4 filter for [HDF5.jl](https://github.com/JuliaIO/HDF5.jl) in Julia.
+See the [documentation](https://juliaio.github.io/HDF5.jl/stable/filters/#H5Zlz4.jl)
+
+This implements [HDF5 registered filter id 32004](https://portal.hdfgroup.org/display/support/Filters#Filters-32004).
