+++ ext/CodecZstdExt/README.md
@@ -1,0 +1,8 @@
+# CodecZstdExt
+
+Implements the Zstd filter for [HDF5.jl](https://github.com/JuliaIO/HDF5.jl) in Julia.
+See the [documentation](https://juliaio.github.io/HDF5.jl/stable/filters/#H5Zzstd.jl)
+
+This implements [HDF5 ZStandard Filter 32015](https://portal.hdfgroup.org/display/support/Filters#Filters-32015)
+
+The contents of this package is derived from H5Zzstd.jl
